requirements.txt
# Core API and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0

# Data processing
pandas>=2.0.0
xarray>=2023.1.0
netCDF4>=1.6.0
pyarrow>=12.0.0
numpy>=1.24.0

# Geospatial (optional but recommended)
geopandas>=0.13.0
shapely>=2.0.0

# ARGO data handling
argopy>=0.1.13

# Dashboard and visualization
streamlit>=1.28.0
plotly>=5.15.0
pydeck>=0.8.0

# Machine Learning and Embeddings
sentence-transformers>=2.2.0
chromadb>=0.4.0
torch>=2.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Optional: OpenAI integration
openai>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Windows-specific recommendations
# If you encounter issues installing some packages, try:
# pip install --only-binary=all package_name
# or use conda instead of pip for scientific packages:
# conda install -c conda-forge xarray netcdf4 pandas